+++ mimesis/providers/date.py
@@ -1,11 +1,13 @@
 """Provider of data related to date and time."""
 
 from calendar import monthrange, timegm
-from datetime import date, datetime, time
+from datetime import date, datetime, time, timedelta
+from typing import List, Optional, Union
 
+from mimesis.compat import pytz
 from mimesis.data import GMT_OFFSETS, ROMAN_NUMS, TIMEZONES
 from mimesis.providers.base import BaseDataProvider
-from mimesis.typing import DateTime, Timestamp
+from mimesis.typing import Date, DateTime, Time
 from mimesis.utils import pull
 
 __all__ = ['Datetime']
@@ -22,15 +24,53 @@ class Datetime(BaseDataProvider):
         super().__init__(*args, **kwargs)
         self._data = pull('datetime.json', self.locale)
 
+    @staticmethod
+    def bulk_create_datetimes(date_start: DateTime,
+                              date_end: DateTime, **kwargs) -> List[DateTime]:
+        """Bulk create datetime objects.
+
+        This method creates list of datetime objects from
+        ``date_start`` to ``date_end``.
+
+        You can use the following keyword arguments:
+
+        * ``days``
+        * ``hours``
+        * ``minutes``
+        * ``seconds``
+        * ``microseconds``
+
+        See datetime module documentation for more:
+        https://docs.python.org/3.7/library/datetime.html#timedelta-objects
+
+
+        :param date_start: Begin of the range.
+        :param date_end: End of the range.
+        :param kwargs: Keyword arguments for datetime.timedelta
+        :return: List of datetime objects
+        :raises: ValueError: When ``date_start``/``date_end`` not passed and
+            when ``date_start`` larger than ``date_end``.
+        """
+        dt_objects = []
+
+        if not date_start and not date_end:
+            raise ValueError('You must pass date_start and date_end')
+
+        if date_end < date_start:
+            raise ValueError('date_start can not be larger than date_end')
+
+        while date_start <= date_end:
+            date_start += timedelta(**kwargs)
+            dt_objects.append(date_start)
+
+        return dt_objects
+
     def week_date(self, start: int = 2017, end: int = 2018) -> str:
         """Get week number with year.
 
         :param start: From start.
         :param end: To end.
         :return: Week number.
-
-        :Example:
-            2017-W32
         """
         year = self.year(start, end)
         week = self.random.randint(1, 52)
@@ -44,9 +84,6 @@ class Datetime(BaseDataProvider):
 
         :param abbr: Abbreviated day name.
         :return: Day of the week.
-
-        :Example:
-            Wednesday
         """
         key = 'abbr' if abbr else 'name'
         days = self._data['day'].get(key)
@@ -57,24 +94,17 @@ class Datetime(BaseDataProvider):
 
         :param abbr: Abbreviated month name.
         :return: Month name.
-
-        :Example:
-            January
         """
         key = 'abbr' if abbr else 'name'
         months = self._data['month'].get(key)
         return self.random.choice(months)
 
-    def year(self, minimum: int = 1990,
-             maximum: int = 2050) -> int:
+    def year(self, minimum: int = 1990, maximum: int = 2050) -> int:
         """Generate a random year.
 
         :param minimum: Minimum value.
         :param maximum: Maximum value.
         :return: Year.
-
-        :Example:
-            2023
         """
         return self.random.randint(minimum, maximum)
 
@@ -82,9 +112,6 @@ class Datetime(BaseDataProvider):
         """Get a random century.
 
         :return: Century.
-
-        :Example:
-            XXI
         """
         return self.random.choice(ROMAN_NUMS)
 
@@ -92,67 +119,68 @@ class Datetime(BaseDataProvider):
         """Get a random periodicity string.
 
         :return: Periodicity.
-
-        :Example:
-            Never.
         """
         periodicity = self._data['periodicity']
         return self.random.choice(periodicity)
 
-    def date(self, start: int = 2000,
-             end: int = 2035, fmt: str = '') -> str:
-        """Generate a string representing of random date.
-
-        Date can be automatically formatted for the current
-        locale or specified.
+    def date(self, start: int = 2000, end: int = 2019) -> Date:
+        """Generate random date object.
 
         :param start: Minimum value of year.
         :param end: Maximum value of year.
-        :param fmt: Format string for date.
         :return: Formatted date.
-
-        :Example:
-            08/16/88 (en)
         """
+        year = self.random.randint(start, end)
+        month = self.random.randint(1, 12)
+        day = self.random.randint(1, monthrange(year, month)[1])
+        date_object = date(year, month, day)
+        return date_object
+
+    def formatted_date(self, fmt: str = '', **kwargs) -> str:
+        """Generate random date as string.
+
+        :param fmt: Format of date, if None then use standard
+            for accepted for current locale
+        :param kwargs: Keyword arguments for :meth:`~Datetime.date()`
+        :return: Formatted date.
+        """
+        date_obj = self.date(**kwargs)
+
         if not fmt:
             fmt = self._data['formats'].get('date')
 
-        year = self.random.randint(start, end)
-        month = self.random.randint(1, 12)
-        d = date(year, month, self.random.randint(
-            1, monthrange(year, month)[1]))
-        return d.strftime(fmt)
+        return date_obj.strftime(fmt)
 
-    def time(self, fmt: str = '') -> str:
-        """Generate a random time.
+    def time(self) -> Time:
+        """Generate a random time object.
 
-        Time can be automatically formatted for the current
-        locale or specified.
-
-        :param fmt: Format of time.
-        :return: Time.
-
-        :Example:
-            21:30:00
+        :return: ``datetime.time`` object.
         """
-        if not fmt:
-            fmt = self._data['formats'].get('time')
-
-        t = time(
+        random_time = time(
             self.random.randint(0, 23),
             self.random.randint(0, 59),
             self.random.randint(0, 59),
             self.random.randint(0, 999999),
         )
-        return t.strftime(fmt)
+        return random_time
+
+    def formatted_time(self, fmt: str = '') -> str:
+        """Generate string formatted time.
+
+        :param fmt: Format of time, if None then use standard
+            for accepted for current locale
+        :return: String formatted time.
+        """
+        time_obj = self.time()
+
+        if not fmt:
+            fmt = self._data['formats'].get('time')
+        return time_obj.strftime(fmt)
 
     def day_of_month(self) -> int:
         """Generate a random day of month, from 1 to 31.
 
         :return: Random value from 1 to 31.
-
-        :Example:
-            23
         """
         return self.random.randint(1, 31)
 
@@ -160,9 +188,6 @@ class Datetime(BaseDataProvider):
         """Get a random timezone.
 
         :return: Timezone.
-
-        :Example:
-            Europe/Paris
         """
         return self.random.choice(TIMEZONES)
 
@@ -170,46 +195,52 @@ class Datetime(BaseDataProvider):
         """Get a random GMT offset value.
 
         :return: GMT Offset.
-
-        :Example:
-            'UTC +03:00'
         """
         return self.random.choice(GMT_OFFSETS)
 
-    def datetime(self, humanized: bool = False, **kwargs) -> DateTime:
+    def datetime(self, start: int = 2000, end: int = 2035,
+                 timezone: Optional[str] = None) -> DateTime:
         """Generate random datetime.
 
-        :param humanized: Readable representation.
-        :param kwargs: Keyword arguments (start, end).
-        :return: Datetime.
-        :rtype: datetime.datetime
-
-        :Example:
-            March, 24 2002
+        :param start: Minimum value of year.
+        :param end: Maximum value of year.
+        :param timezone: Set custom timezone (pytz required).
+        :return: Datetime
         """
-        fmt = '%Y-%m-%d %H:%M:%S'
-        dt_str = '{date} {time}'.format(
-            date=self.date(fmt='%Y-%m-%d', **kwargs),
+        datetime_obj = datetime.combine(
+            date=self.date(start, end),
             time=self.time(),
         )
+        if timezone:
+            if not pytz:
+                raise ImportError('Timezones are supported only with pytz')
+            tz = pytz.timezone(timezone)
+            datetime_obj = tz.localize(datetime_obj)
 
-        dt = datetime.strptime(dt_str, fmt)
+        return datetime_obj
 
-        if humanized:
-            return dt.strftime('%B, %d %Y')
+    def formatted_datetime(self, fmt: str = '', **kwargs) -> str:
+        """Generate datetime string in human readable format.
 
-        return dt
+        :param fmt: Custom format (default is format for current locale)
+        :param kwargs: Keyword arguments for :meth:`~Datetime.datetime()`
+        :return: Formatted datetime string.
+        """
+        dt_obj = self.datetime(**kwargs)
 
-    def timestamp(self, posix: bool = True, **kwargs) -> Timestamp:
+        if not fmt:
+            date_fmt = self._data['formats'].get('date')
+            time_fmt = self._data['formats'].get('time')
+            fmt = '{} {}'.format(date_fmt, time_fmt)
+
+        return dt_obj.strftime(fmt)
+
+    def timestamp(self, posix: bool = True, **kwargs) -> Union[str, int]:
         """Generate random timestamp.
 
         :param posix: POSIX time.
-        :param kwargs: Keyword arguments (start, end).
+        :param kwargs: Kwargs for :meth:`~Datetime.datetime()`.
         :return: Timestamp.
-        :rtype: str or int
-
-        :Example:
-            2018-01-02T06:19:19Z
         """
         stamp = self.datetime(**kwargs)