+++ mimesis/typing.py
@@ -9,13 +9,12 @@ __all__ = [
     'Bytes',
     'DateTime',
     'Timestamp',
+    'Time',
+    'Date',
     'Seed',
 ]
 
-JSON = Union[
-    Dict[Any, Any],
-    Any,
-]
+JSON = Dict[str, Any]
 
 _StrOrInt = Union[str, int]
 
@@ -23,7 +22,11 @@ Size = _StrOrInt
 
 Bytes = bytes
 
-DateTime = Union[datetime.datetime, Any]
+DateTime = datetime.datetime
+
+Time = datetime.time
+
+Date = datetime.date
 
 Timestamp = _StrOrInt